[#36] Decompose parser sources
